Replace clients file by Client model
