+++ .gitignore
@@ -103,6 +103,5 @@ dist
 # TernJS port file
 .tern-port
 
-/clients.yml
 /elus.json
 /admins.json
