--- clients.yml.sample
@@ -1,4 +1,0 @@
-- nom: iAdresses
-  organisme: ACME SARL
-  email: iadresses@acme.ltd
-  token: xxxxxxxxxxxxxxx
