+++ lib/habilitations/__tests__/routes.js
@@ -7,11 +7,6 @@ const {habilitationsRoutes} = require('../routes')
 
 let mongod
 
-const fakeClient = {
-  name: 'ACME',
-  token: 'foobar'
-}
-
 test.before('start server', async () => {
   mongod = await MongoMemoryServer.create()
   await mongo.connect(mongod.getUri())
@@ -39,7 +34,15 @@ test('fetch habilitation', async t => {
   const habilitation = {_id, codeCommune: '27115'}
   await mongo.db.collection('habilitations').insertOne(habilitation)
 
-  const server = await getApp({clients: [fakeClient]})
+  const clientId = new mongo.ObjectId()
+  const client = {
+    _id: clientId,
+    name: 'ACME',
+    token: 'foobar'
+  }
+  await mongo.db.collection('clients').insertOne(client)
+
+  const server = await getApp()
   const res = await request(server)
     .get(`/habilitations/${_id}`)
     .set('Authorization', 'Token foobar')
@@ -49,7 +52,13 @@ test('fetch habilitation', async t => {
 })
 
 test('basic habilitation', async t => {
-  const server = await getApp({clients: [fakeClient]})
+  const clientId = new mongo.ObjectId()
+  await mongo.db.collection('clients').insertOne({
+    _id: clientId,
+    name: 'ACME',
+    token: 'foobar'
+  })
+  const server = await getApp()
 
   const res1 = await request(server)
     .post('/communes/27115/habilitations')
