+++ lib/habilitations/routes.js
@@ -7,9 +7,8 @@ const {omit} = require('lodash')
 const mongo = require('../util/mongo')
 const errorHandler = require('../util/error-handler')
 const {getCommune} = require('../util/cog')
-const {ensureCodeCommune, ensureCommuneActuelle} = require('../util/middlewares')
+const {authClient, ensureCodeCommune, ensureCommuneActuelle} = require('../util/middlewares')
 const w = require('../util/w')
-const {createAuthClient} = require('../clients')
 const {sendMail} = require('../util/sendmail')
 const {createHabilitation, askHabilitation, acceptHabilitation, fetchHabilitation, rejectHabilitation} = require('./model')
 const createPinCodeEmail = require('./strategies/email/pin-code-template')
@@ -18,9 +17,8 @@ const {createFranceConnectStrategy, getMandatCommune} = require('./strategies/fr
 
 const DEMO_MODE = process.env.DEMO_MODE === '1'
 
-async function habilitationsRoutes(params = {}) {
+async function habilitationsRoutes() {
   const app = new express.Router()
-  const authClient = createAuthClient({clients: params.clients})
 
   app.use(express.json())
