+++ lib/revisions/__tests__/integration.js
@@ -9,11 +9,6 @@ const {revisionsRoutes} = require('../routes')
 
 let mongod
 
-const fakeClient = {
-  name: 'ACME',
-  token: 'foobar'
-}
-
 test.before('start server', async () => {
   mongod = await MongoMemoryServer.create()
   await mongo.connect(mongod.getUri())
@@ -38,7 +33,15 @@ async function getApp(params) {
 }
 
 test('authentication / success', async t => {
-  const server = await getApp({clients: [fakeClient]})
+  const server = await getApp()
+
+  const clientId = new mongo.ObjectId()
+  const client = {
+    _id: clientId,
+    name: 'ACME',
+    token: 'foobar'
+  }
+  await mongo.db.collection('clients').insertOne(client)
   const res = await request(server)
     .get('/me')
     .set('Authorization', 'Token foobar')
@@ -48,7 +51,16 @@ test('authentication / success', async t => {
 })
 
 test('authentication / fail', async t => {
-  const server = await getApp({clients: [fakeClient]})
+  const server = await getApp()
+
+  const clientId = new mongo.ObjectId()
+  const client = {
+    _id: clientId,
+    name: 'ACME',
+    token: 'foobar'
+  }
+  await mongo.db.collection('clients').insertOne(client)
+
   await request(server)
     .get('/me')
     .set('Authorization', 'Token ohnooo')
@@ -58,7 +70,15 @@ test('authentication / fail', async t => {
 })
 
 test.serial('basic revision', async t => {
-  const server = await getApp({clients: [fakeClient]})
+  const server = await getApp()
+
+  const clientId = new mongo.ObjectId()
+  const client = {
+    _id: clientId,
+    name: 'ACME',
+    token: 'foobar'
+  }
+  await mongo.db.collection('clients').insertOne(client)
 
   const res1 = await request(server)
     .post('/communes/31591/revisions')
