+++ lib/revisions/__tests__/routes.js
@@ -7,11 +7,6 @@ const {revisionsRoutes} = require('../routes')
 
 let mongod
 
-const fakeClient = {
-  name: 'ACME',
-  token: 'foobar'
-}
-
 test.before('start server', async () => {
   mongod = await MongoMemoryServer.create()
   await mongo.connect(mongod.getUri())
@@ -43,8 +38,15 @@ test.serial('publish revision / without habilitation', async t => {
     codeCommune: '27115',
     client: {name: 'ACME'}
   })
+  const clientId = new mongo.ObjectId()
+  const client = {
+    _id: clientId,
+    name: 'ACME',
+    token: 'foobar'
+  }
+  await mongo.db.collection('clients').insertOne(client)
 
-  const server = await getApp({clients: [fakeClient]})
+  const server = await getApp()
   const res = await request(server)
     .post(`/revisions/${_id}/publish`)
     .set('Authorization', 'Token foobar')
@@ -74,8 +76,15 @@ test.serial('publish revision / with habilitation', async t => {
     codeCommune: '27115',
     client: {name: 'ACME'}
   })
+  const clientId = new mongo.ObjectId()
+  const client = {
+    _id: clientId,
+    name: 'ACME',
+    token: 'foobar'
+  }
+  await mongo.db.collection('clients').insertOne(client)
 
-  const server = await getApp({clients: [fakeClient]})
+  const server = await getApp()
   const res = await request(server)
     .post(`/revisions/${revisionId}/publish`)
     .set('Authorization', 'Token foobar')
