+++ lib/revisions/routes.js
@@ -2,16 +2,14 @@ const express = require('express')
 const {isString, isPlainObject} = require('lodash')
 const createError = require('http-errors')
 const errorHandler = require('../util/error-handler')
-const {ensureCodeCommune, ensureCommuneActuelle} = require('../util/middlewares')
+const {authClient, ensureCodeCommune, ensureCommuneActuelle} = require('../util/middlewares')
 const w = require('../util/w')
 const rawBodyParser = require('../util/raw-body-parser')
-const {createAuthClient} = require('../clients')
 const {validateBAL} = require('./validate-bal')
 const {fetchRevision, getCurrentRevision, getRevisionsByCommune, createRevision, setFile, getFiles, getFileData, publishRevision, computeRevision, getCurrentRevisions, getRelatedHabilitation} = require('./model')
 
-async function revisionsRoutes(params = {}) {
+async function revisionsRoutes() {
   const app = new express.Router()
-  const authClient = createAuthClient({clients: params.clients})
 
   app.use(express.json())
